submission/requirements.txt
pandas
numpy
scikit-learn
matplotlib
google-generativeai